Include the current buffer in results
Instead of preventing the current buffer from being shown in the
results, score it lower.  This way, it can still be searched for.

Any scoring bonus that is a direct consequence of the buffer being
current is removed.  This includes recency, frecency, open, and
proximity. This should prevent the file from being scored so highly that
it clutters the results, but still allows the file to be searched for.

If the current file is selected, no "learning" occurs, and it is treated
as if the search dialog were canceled.

Fixes #10