Highlight entries properly
This is a re-engineering of highlighting that uses the appropriate
highlighter for the selected match algorithm.  The highlighting has to
be done in the display function because we have to have access to
the matcher actually matched on, which is different from what we end up
displaying.  This is because the matchers are looking at a full path,
but we break up the path to show the filename first.