Open buffer indicators
Show indicator if the buffer is currently open, and allow buffers to be
closed from the telescope window by pressing Ctrl-c.  (This mapping
should be made configurable at some point).

For some reason, we have to hard-code the display width of the
indicators, as vim.fn.strdisplaywidth doesn't give the right value for
the width of these characters.  (vim.fn.strlen does, but that function
doesn't seem to provide the right values elsewhere)