Fix entry duplication
Only use db results directly if there are enough entries to fill the
available slots. This is a performance optimization so that we can call
process_complete more quickly than waiting for a full scan and rank of
all files, which should allow the results to pop up quickly in the cases
where the user doesn't need to type anything.